520 Address review feedback
- Use `long` for `WorkflowTemplatePage.totalElements` to match other page DTOs and Spring `Page#getTotalElements()`
- Fix duplicated `@author` tag in `AutomationWorkflowProjectDTO`

+++ server/libs/platform/platform-github-proxy-client/src/main/java/com/bytechef/platform/githubproxy/client/model/WorkflowTemplatePage.java
@@ -24,5 +24,5 @@ import java.util.List;
  */
 @SuppressFBWarnings("EI")
 public record WorkflowTemplatePage(
-    List<WorkflowTemplateSummary> content, int page, int size, int totalElements, int totalPages) {
+    List<WorkflowTemplateSummary> content, int page, int size, long totalElements, int totalPages) {
 }
